% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjR2.R
\name{adjR2}
\alias{adjR2}
\title{Adjusted coefficient of determination}
\usage{
adjR2(r2, n, p)
}
\arguments{
\item{r2}{Numeric vector with the coefficient of determination to be
adjusted.}

\item{n}{Numeric vector providing the number of observations used to fit the
multiple linear regression model.}

\item{p}{Numeric vector providing the number of parameters included in the
multiple linear regression model.}
}
\value{
A numeric vector with the adjusted coefficient of determination.
}
\description{
Calculates the adjusted coefficient of determination of a multiple linear
regression model.
}
\details{
Details will be added later.
}
\examples{
adjR2(r2 = 0.95, n = 100, p = 80)
}
\references{
Coefficient of determination. Wikipedia, The Free Encyclopedia.
Available at https://en.wikipedia.org/wiki/Coefficient_of_determination.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\keyword{misc}
