\name{cdfTable}
\alias{cdfTable}
\title{
Table with descriptive statistics of an estimated cummulative distribution
function (CDF)
}
\description{
  This function returns a table containing the descriptive statistics of the
  CDF of a set of continuous variables. TeX code is printed to copy and paste in 
  a document.
}
\usage{
cdfTable(x, type = "xy", rounding = 0, tex = FALSE, data.frame = FALSE)
}
\arguments{
  \item{x}{
    Object with the estimated CDF of the set of continuous variables. The
    resulting object of \code{cont.analysis} of package \bold{spsurvey}.
  }
  \item{type}{
    Type of data under analysis. Defaults to \code{type = "xy"}. See
    \sQuote{Details}.
  }
  \item{rounding}{
    Rounding level of the data in the output table. Defaults to
    \code{rounding = 0}.
  }
  \item{tex}{Logical for creating TeX code. Defaults to \code{tex = FALSE}.
  }
  \item{data.frame}{Logical for returning a data.frame object. Defaults to
     \code{data.frame = FALSE}.}
}
\details{
  Summary statistics included in the table (estimated population mean and
  standard deviation) are obtained from the resulting object of 
  \code{cont.analysis} by internaly using the function \code{cdfStats}.
  
  There are two types of data that can be submitted to function \code{cdfTable}.
  The first (\code{type = "xy"}) is composed by two instances ('x' and 'y') and
  is produced during horizontal (positional) validation exercises (validation in
  the geographic space). Thus, 'x' and 'y' represent, respectively, the 
  horizontal displacement (error) in 'x' and 'y' coordinates.
  
  The second type of data (\code{type = "z"}) is composed by only one instance
  (\sQuote{z}) and is generated by vertical validation exercises (validation in
  the attribute space). Thus, \sQuote{z} represents the vertical displacement
  (error) of the attribute \sQuote{z} being measured.
}
\value{
  Returned value depends on how arguments \code{type} and \code{tex} are set.
\item{\code{type}}{
  If \code{type = "xy"}, then the function returns a table with
  estimated population mean and standard deviation of error statistics
  for 'x' and 'y' coordinates. These error statistics include the mean
  error, mean absolute error, and mean square error. It also returns
  the  estimated mean and mean square error vector (module), and the
  estimated mean azimuth. The number of ground control points used to
  make the estimates is printed by default.
  
  If \code{type = "z"}, then the function returns a table with
  estimated population mean and standard deviation of error statistics
  for 'z', the attribute under analysis. These error statistics include
  the mean error, mean absolute error, and mean square error. The
  number of ground control points used to make the estimates is printed
  by default.}
\item{\code{tex}}{
  If \code{tex = TRUE}, them the function prints the TeX code for the
  table defined by the argument \code{type}. Otherwise the TeX code is
  not generated.}
}
\references{
Kincaid, T. M. and Olsen, A. R. (2013).
  spsurvey: Spatial Survey Design and Analysis.
  R package version 2.6. URL: <\url{http://www.epa.gov/nheerl/arm/}>.
}
\author{
Alessandro Samuel-Rosa <\email{alessandrosamuelrosa@gmail.com}>
}
\seealso{
  \code{\link[pedometrics]{cdfStats}},
  \code{\link[spsurvey]{cont.analysis}}.
}
\examples{
\dontrun{
## Estimate the CDF
my.cdf <- cont.analysis(spsurvey.obj = my.spsurvey)

## Print table and TeX code
cdfTable(my.cdf)
}
}
\keyword{methods}
\keyword{print}
% End!