% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkGMU.R
\name{checkGMU}
\alias{checkGMU}
\title{Evaluation of geostatistical models of uncertainty}
\usage{
checkGMU(observed, simulated, pi = seq(0.01, 0.99, 0.01), symmetric = TRUE,
  plotit = TRUE)
}
\arguments{
\item{observed}{Vector of observed values at the validation points. See
\sQuote{Details} for more information.}

\item{simulated}{Data frame or matrix with simulated values (columns) for
each validation point (rows). See \sQuote{Details} for more information.}

\item{pi}{Vector defining the width of the series of probability intervals.
Defaults to \code{pi = seq(0.01, 0.99, 0.01)}. See \sQuote{Details} for more
information.}

\item{symmetric}{Logical for choosing the type of probability interval.
Defaults to \code{symmetric = TRUE}. See \sQuote{Details} for more
information.}

\item{plotit}{Logical for plotting the results. Defaults to
\code{plotit = TRUE}.}
}
\description{
Evaluate the local quality of a geostatistical model of uncertainty
(GMU) using summary measures and graphical displays.
}
\details{
There is no standard way of evaluating the local quality of a GMU. The
collection of summary measures and graphical displays presented here is far
from being comprehensive, nor is it intended to be the \sQuote{best}
selection of existing summary measures and graphical displays. My choice
is purely based on the literature -- I have tried to include here those
summary measures and graphical displays that are most commonly used to
evaluate the local quality of GMUs.

Understanding a few definitions is needed to correctly use the collection of
summary measures and graphical displays presented here. I try to give a
brief overview of them in the following lines.

\subsection{Error statistics}{
Error statistics measure how well the GMU predicts the measured values at the
validation points. Four error statistics are presented:

\describe{
\item{Mean error (ME)}{
Measures the bias of the predictions of the GMU, being defined as the mean of
the differences between the average of the simulated values and the observed
values, i.e. the average of all simulations is taken as the predicted value.
}
\item{Mean squared error (MSE)}{
Measures the accuracy of the predictions of the GMU, being defined as the
mean of the squared differences between the average of the simulated values
and the observed values.
}
\item{Scaled root mean squared error (SRMSE)}{
Measures how well the GMU estimate of the prediction error variance (PEV)
approximates the observed prediction error variance, where the first is
given by the variance of the simulated values, while the second is given by
the squared differences between the average of the simulated values, i.e.
the squared error (SE). The SRMSE is computed as the average of SE / PEV,
where SRMSE > 1 indicates underestimation, while SRMSE < 1 indicates
overestimation.
}
\item{Pearson correlation coefficient}{
Measures how close the GMU predictions are to the observed values. A scatter
plot of the observed values versus the average of the simulated values
can be used to checking for possible unwanted outliers and non-linearities.
The square of the Pearson correlation coefficient measures the fraction
of the overall spread of observed values that is explained by the GMU,
that is, the amount of variance explained (AVE), also known as coefﬁcient
of determination or ratio of scatter.
}
}
}
\subsection{Coverage probabilities}{
The coverage probability of an interval is given by the number of times that
that interval contains its parameter over infinite, independent, and
identical replications of an experiment. For example, the interval defined
by the lower and upper quartiles (0.25, 0.75) of a variable contains, by
definition, 0.5 of the values of that variable -- that is the nominal
coverage probability of that interval. Consider now a Gaussian distributed
variable with mean equal to zero and variance equal to one: knowledge of the
mean and variance allows us to compute the exact values of the lower and
upper quartiles. If we generate a Gaussian distributed \emph{random}
variable with the same mean and variance, about 0.5 of its values will fall
in the quartilic interval of our Gaussian variable. If we continue
generating Gaussian distributed \emph{random} variables with the same mean
and variance, on average, 0.5 of the values will fall in that interval.

Coverage probabilities are very useful to evaluate the local quality of a
GMU: the closer the observed coverage probabilities of a sequence of
probability intervals (PI) are to the nominal coverage probabilities of
those PIs, the better the modelling of the local uncertainty. Two types of
PIs can be used here: symmetric, median-centred PIs, and left-bounded PIs.
Papritz & Dubois (1999) recommend using left-bounded PIs because they are
better at evidencing deviations for both large and small PIs. The authors
also point that the coverage probabilities of the symmetric, median-centred
PIs can be read from the coverage probability plots produced using
left-bounded PIs.

Deutsch (1997) proposed three summary measures of the coverage
probabilities to assess the local \emph{goodness} of a GMU: accuracy ($A$),
precision ($P$), and goodness ($G$). According to Deutsch (1997), a GMU can
be considered \dQuote{good} if it is both accurate and precise. Although
easy to compute, these measures seem not to have been explored by many
geostatiticians, except for the studies developed by Pierre Goovaerts and
his later software implementation (Goovaerts, 2009). Richmond (2001)
suggests that they should not be used as the only measures of the local
quality of a GMU.

\describe{
\item{Accuracy}{
An accurate GMU is that for which the proportion $p^*$ of true values
falling within the symmetric $p$ PI is equal to or larger than $p$, that is,
when $p^* \geq p$. Thus, a GMU will be more accurate when all points in the
coverage probability plot are on or above the 1:1 line. The range of $A$
goes from 0 (lest accurate) to 1 (most accurate).
}
\item{Precision}{
The \emph{precision}, $P$, is defined only for an accurate GMU, and measures
how close $p^*$ is to $p$. The range of $P$ goes from 0 (lest precise) to 1
(most precise). Thus, a GMU will be more accurate when all points in the
PI-width plot are on or above the 1:1 line.
}
\item{Goodness}{
The \emph{goodness}, $G$, is a measure of the departure of the points from
the 1:1 line in the coverage probability plot. $G$ ranges from 0 (minimum
goodness) to 1 (maximum goodness), the maximum $G$ being achieved when
$p^* = p$, that is, all points in both coverage probability and interval
width plots are exactly on the 1:1 line.
}
}
}
}
\examples{
set.seed(2001)
observed <- round(rnorm(100), 3)
simulated <- t(sapply(1:length(observed),
                      function (i) round(rnorm(100), 3)))
resa <- checkGMU(observed, simulated, symmetric = T)
resb <- checkGMU(observed, simulated, symmetric = F)
round(resa$stats, 3)
round(resb$stats, 3)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Deutsch, C. Direct assessment of local accuracy and precision. Baafi, E. Y.
& Schofield, N. A. (Eds.) \emph{Geostatistics Wollongong '96}. Dordrecht:
Kinwer Academic Pubiishers, v. I, p. 115-125, 1997.

Papritz, A. & Dubois, J. R. Mapping heavy metals in soil by (non-)linear
kriging: an empirical validation. Gómez-Hernández, J.; Soares, A. &
Froidevaux, R. (Eds.) \emph{geoENV II -- Geostatistics for Environmental
Applications}. Springer, p. 429-440, 1999.

Goovaerts, P. Geostatistical modelling of uncertainty in soil science.
\emph{Geoderma}. v. 103, p. 3 - 26, 2001.

Goovaerts, P. AUTO-IK: a 2D indicator kriging program for the automated
non-parametric modeling of local uncertainty in earth sciences.
\emph{Computers & Geosciences}. v. 35, p. 1255-1270, 2009.

Richmond, A. J. Maximum profitability with minimum risk and effort. Xie, H.;
Wang, Y. & Jiang, Y. (Eds.) \emph{Proceedings 29th APCOM}. Lisse: A. A.
Balkema, p. 45-50, 2001.
}

