% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popStats.R
\name{COV}
\alias{COV}
\title{Population statistics}
\usage{
COV(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\description{
Calculate population statistics (variance and covariance).
}
\note{
These functions will be implemented in C++ using the R-package \pkg{Rcpp}.
}
\examples{
x <- rnorm(10)
y <- rnorm(10)
res <- data.frame(Variance = c(var(x), VAR(x)),
                  Covariance = c(cov(x, y), COV(x, y)),
                  row.names = c("Sample", "Population"))
res
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Rencher, A. C. Methods of multivariate analysis. New York: John Wiley & Sons,
p. 708, 2002.
}
\seealso{
\code{\link[stats]{cor}}
}

