% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcpDiff.R
\name{gcpDiff}
\alias{gcpDiff}
\title{Difference on xyz coordinates between ground control points}
\usage{
gcpDiff(measured, predicted, type = "xy", aggregate = FALSE, rounding = 0)
}
\arguments{
\item{measured}{Object of class \code{SpatialPointsDataFrame}
with the reference GCP. A column named \sQuote{siteID} giving case names is
mandatory. See \sQuote{Details}, item \sQuote{Type of data}.}

\item{predicted}{An object of class
\code{SpatialPointsDataFrame} with the point data being
validated. A column named \sQuote{siteID} giving case names is mandatory.
See \sQuote{Details}, item \sQuote{Type of data}.}

\item{type}{Type of data under analysis. Defaults to \code{type = "xy"}.
\sQuote{Details}, item \sQuote{Type of data}.}

\item{aggregate}{Logical for aggregating the data when it comes from cluster
sampling. Used only when \code{type = "z"}. Defaults to \code{aggregate =
FALSE}. See \sQuote{Details}, item \sQuote{Data aggregation}.}

\item{rounding}{Rounding level of the data in the output data frame.}
}
\value{
An object of class \code{data.frame} ready to be used to feed the
argument \code{data.cont} when creating a \code{spsurvey.analysis} object.
}
\description{
This function estimates the difference, absolute difference, and squared
difference on x, y and z coordinates of two sets of ground control points
(GCP). It also estimates the module (difference vector), its square and
azimuth. The result is a data frame ready to be used to define a object of
class \code{spsurvey.object}.
}
\details{
\subsection{Type of data}{
Two types of validation data that can be submitted to function
\code{gcpDiff()}: those coming from horizontal (positional) validation
exercises (\code{type = "xy"}), and those coming from vertical validation
exercises (\code{type = "z"}).

Horizontal (positional) validation exercises compare the position of
\code{measured} point data with the position of \code{predicted} point data.
Horizontal displacement (error) is measured in both \sQuote{x} and
\sQuote{y} coordinates, and is used to calculate the error vector (module)
and its azimuth.  Both objects \code{measured} and \code{predicted} used
with function \code{gcpDiff()} must be of class
\code{SpatialPointsDataFrame}. They must have at least one column named
\sQuote{siteID} giving the identification of every case. Matching of case
IDs is mandatory. Other columns are discarded.

Vertical validation exercises are interested in comparing the
\code{measured} value of a variable at a given location with that
\code{predicted} by some model. In this case, error statistics are
calculated only for the the vertical displacement (error) in the \sQuote{z}
coordinate. Both objects \code{measured} and \code{predicted} used with
function \code{gcpDiff()} must be of class \code{SpatialPointsDataFrame}.
They also must have a column named \sQuote{siteID} giving the identification
of every case. Again, matching of case IDs is mandatory. However, both
objects must have a column named \sQuote{z} which contains the values of the
\sQuote{z} coordinate. Other columns are discarded.
}
\subsection{Data aggregation}{
Validation is sometimes performed using cluster or transect sampling. Before
estimation of error statistics, the data needs to be aggregated by cluster
or transect. The function \code{gcpDiff()} aggregates validation data of
\code{type = "z"} calculating the mean value per cluster. Thus, aggregation
can only be properly done if the \sQuote{siteID} column of both objects
\code{measured} and \code{predicted} provides the identification of
clusters.  Setting \code{aggregate = TRUE} will return aggregated estimates
of error statistics. If the data has been aggregated beforehand, the
parameter \code{aggregate} can be set to \code{FALSE}.
}
\subsection{Case matching}{
There are circumstances in which the number of cases in the object
\code{measured} is larger than that in the object \code{predicted}. The
function \code{gcpDiff()} compares the number of cases in both objects and
automatically drops those cases of object \code{measured} that do not match
the cases of object \code{predicted}. However, case matching can only be
done if case IDs are exactly the same for both objects. Otherwise, estimated
error statistics will have no meaning at all.
}
}
\note{
Data of \code{type = "xy"} cannot be submitted to cluster aggregation
in the present version.
}
\examples{

\dontrun{
if (require(spsurvey)) {
## Create an spsurvey.analysis object
my.spsurvey <- 
  spsurvey.analysis(design = coordenadas(my.data),
                    data.cont = delta(ref.data, my.data),
                    popcorrect = TRUE, pcfsize = length(my.data$id),
                    support = rep(1, length(my.data$id)),
                    wgt = rep(1, length(my.data$id)), vartype = "SRS")
}
}

}
\references{
Kincaid, T. M. and Olsen, A. R. (2013). spsurvey: Spatial Survey
Design and Analysis. R package version 2.6. URL:
\url{http://www.epa.gov/nheerl/arm/}.
}
\seealso{
\code{\link[pedometrics]{coordenadas}},
\code{\link[pedometrics]{gcpVector}},
\code{\link[spsurvey]{spsurvey.analysis}}.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\keyword{methods}
