% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedometrics-defunct.R
\name{coordenadas}
\alias{coordenadas}
\alias{febr}
\alias{febr2spdf}
\alias{febr2xlsx}
\alias{standard}
\alias{header}
\alias{cdfPlot}
\alias{cdfStats}
\title{Defunct functions in the \strong{pedometrics} package}
\usage{
coordenadas(...)

cdfPlot(...)

cdfStats(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
No return value, called for side effects.
}
\description{
The functions listed here are no longer part of the \strong{pedometrics} package as they are not
needed (any more).
}
