\name{pedometrics-package}
\alias{pedometrics-package}
\alias{pedometrics}
\docType{package}
\title{Pedometric Tools and Techniques}
\description{
This package contains functions to employ many tools and techniques used in the field of pedometrics (see http://en.wikipedia.org/wiki/Pedometrics for definition of pedometrics). These functions were developed to fulfill the demands created by the PhD research project (2012-2016) entitled Contribution to the Construction of Models for Predicting Soil Properties, developed by Alessandro Samuel-Rosa under the supervision of Dr Lucia HC Anjos (Federal Rural University of Rio de Janeiro, Brazil), Dr Gustavo M Vasques (Embrapa Soils, Brazil), and Dr Gerard BM Heuvelink (ISRIC - World Soil Information, the Netherlands). The project is supported by the CAPES Foundation, Ministry of Education of Brazil, Brasilia - DF 70040-020, Brazil, process ID BEX 11677/13-9. Several functions are simple wrapper functions that extend the functionalities of other functions commonly used for the analysis of pedometric data. It should be noted that changes are likely to occur quite often and using this package as a dependency for other packages is strongly discouraged.
}
\details{
\tabular{ll}{
Package: \tab pedometrics\cr
Type: \tab Package\cr
Version: \tab 0.1-9\cr
Date: \tab 2014-08-1\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Author: Alessandro Samuel-Rosa, Lucia H. C. Anjos, Gustavo M. Vasques, Gerard B. M. Heuvelink

Maintainer: Alessandro Samuel-Rosa <alessandrosamuelrosa@gmail.com>
}
