% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trend.R
\name{trend.terms}
\alias{trend.matrix}
\alias{trend.terms}
\title{Extract spatial trend data}
\usage{
trend.terms(x)

trend.matrix(x)
}
\arguments{
\item{x}{Object of class \code{likfit}.}
}
\description{
Extract spatial trend data from an object of class \code{likfit}.
}
\details{
\code{trend.terms} is similar to \code{\link[stats]{terms}}.

\code{trend.matrix} is similar to \code{\link[stats]{model.frame}}.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\seealso{
\code{\link[geoR]{likfit}}
}

