% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordenadas.R
\name{coordenadas}
\alias{coordenadas}
\title{Prepare object for argument \code{design} of \code{spsurvey.analysis()}}
\usage{
coordenadas(x)
}
\arguments{
\item{x}{Object of class \code{SpatialPointsDataFrame} from which site ID and XY coordinates are to be returned.}
}
\value{
An object of class \code{data.frame} containing three columns with names \code{siteID}, \code{xcoord}, and
\code{ycoord}.
}
\description{
This function returns an object to feed the argument \code{design} when
creating an object of class \code{spsurvey.analysis}.
}
\details{
The argument \code{design} used to create object of class
\code{spsurvey.analysis} requires a series of inputs. However, it can be fed
with data about site ID and coordinates. \code{coordenadas()} returns a data
frame that provides this information, assuming that all other design
variables are provided manually in the arguments list.
}
\examples{

\dontrun{
if (require(spsurvey)) {
  ## Create an spsurvey.analysis object
  my.spsurvey <-
    spsurvey::spsurvey.analysis(
      design = coordenadas(my.data),
      data.cont = delta(ref.data, my.data),
      popcorrect = TRUE, pcfsize = length(my.data$id),
      support = rep(1, length(my.data$id)),
      wgt = rep(1, length(my.data$id)), vartype = "SRS")
}
}

}
\references{
Kincaid, T. M. and Olsen, A. R. (2013). \emph{spsurvey: Spatial Survey Design and Analysis}.
R package version 2.6. URL: \url{https://www.epa.gov/nheerl/arm/}.
}
\seealso{
\code{\link[pedometrics]{gcpDiff}}, \code{\link[spsurvey]{cont.analysis}}.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\keyword{methods}
