% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedometrics-defunct.R
\name{coordenadas}
\alias{coordenadas}
\alias{cdfPlot}
\alias{cdfStats}
\alias{cdfTable}
\alias{gcpDiff}
\alias{trend.terms}
\alias{trend.matrix}
\title{Defunct functions in the \strong{pedometrics} package}
\usage{
coordenadas(...)

cdfPlot(...)

cdfStats(...)

cdfTable(...)

gcpDiff(...)

trend.terms(...)

trend.matrix(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
No return value, called for side effects.
}
\description{
The functions listed here are no longer part of the \strong{pedometrics} package. If you need to use
any of these functions, you can still found at \url{https://github.com/samuel-rosa/ASRtools}.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}\cr
Tony Olsen \email{Olsen.Tony@epa.gov}\cr
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
