% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowMinCpp}
\alias{rowMinCpp}
\title{Return the minimum value in each row of a numeric matrix}
\usage{
rowMinCpp(x)
}
\arguments{
\item{x}{Numeric matrix with two or more rows and/or columns.}
}
\value{
A numeric vector with the minimum value of each row if the matrix.
}
\description{
This function returns the minimum value in each row of a numeric matrix.
}
\details{
This function is implemented in C++ to speed-up the computation time for
large matrices.
}
\examples{
x <- matrix(rnorm(20), nrow = 5)
rowMinCpp(x)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\seealso{
\code{\link[matrixStats]{rowMins}}
}
\keyword{univar}

