% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmICP.R
\name{vgmICP}
\alias{vgmICP}
\title{Initial covariance parameters (ICP)}
\usage{
vgmICP(z, coords, lags, max.dist = Inf, method = "a", min.npairs = 30,
  model, nu, plotit = FALSE, ...)
}
\arguments{
\item{z}{Numeric vector with the values of the response variable.}

\item{coords}{Data frame or matrix with the projected x- and y-coordinates.}

\item{lags}{Numeric scalar defining a the width of the lag-distance classes,
or a numeric vector with the upper bounds the lag-distance classes.}

\item{max.dist}{Positive numeric defining the maximum distance up to which
lag-distance classes should be computed. Defaults to \code{max.dist = Inf}.}

\item{method}{Character keyword defining the method used for guessing the
initial covariance parameters. Defauls to \code{method = "a"}. See 
\sQuote{Details} for more information.}

\item{min.npairs}{Positive integer defining the minimum number of 
point-pairs required so that a lag-distance class is used for guessing the
initial covariance parameters. Defaults to \code{min.npairs = 30}.}

\item{model}{Character keyword defining the variogram model that will be
fitted to the data. Currently, most basic variogram models provided by the
\pkg{RandomFields}-package are accepted. See \sQuote{Details} of 
\code{\link[RandomFields]{RMmodel}}.}

\item{nu}{Smoothness parameter \eqn{\nu} of the Whittle-Matérn model. See 
\code{\link[RandomFields]{RMmodel}}.}

\item{plotit}{Should the guessed initial covariance parameters be ploted
along with the sample variogram? Defaults to \code{plotit = FALSE}.}

\item{...}{Further alguments passed to \code{\link[georob]{georob}}.}
}
\description{
Guess the initial values for the covariance parameters.
}

