% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmSCV.R
\name{vgmSCV}
\alias{vgmSCV}
\title{Spatially correlated variance}
\usage{
vgmSCV(obj, digits = 4)
}
\arguments{
\item{obj}{Variogram model fitted with available function in geostatistical
packages such as \pkg{gstat}, \pkg{geoR}, and \pkg{georob}.}

\item{digits}{Integer indicating the number of decimal places to be used.}
}
\value{
Numeric value indicating the proportion of the variance that is 
spatially correlated.
}
\description{
Compute the proportion of the variance that is spatially correlated.
}
\author{
Alessandro Samuel-Rosa <\email{alessandrosamuelrosa@gmail.com}>
}
\seealso{
\code{\link[pedometrics]{vgmLags}}
}
\concept{
variogram
}

