% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmSCV.R
\name{vgmSCV}
\alias{vgmSCV}
\alias{vgmSCV.variomodel}
\alias{vgmSCV.variogramModel}
\alias{vgmSCV.georob}
\title{Spatially correlated variance (SCV)}
\usage{
\method{vgmSCV}{variomodel}(obj, digits = 4)

\method{vgmSCV}{variogramModel}(obj, digits = 4)

\method{vgmSCV}{georob}(obj, digits = 4)
}
\arguments{
\item{obj}{Variogram model fitted with available function in geostatistical
packages such as \pkg{gstat}, \pkg{geoR}, and \pkg{georob}.}

\item{digits}{Integer indicating the number of decimal places to be used.}
}
\value{
Numeric value indicating the proportion of the variance that is
spatially correlated.
}
\description{
Compute the proportion of the variance that is spatially correlated.
}
\seealso{
\code{\link[pedometrics]{vgmLags}}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\concept{variogram}
